Compile fix
svn path=/trunk/quanta/quanta/parts/kfilereplace/; revision=237021

+++ filelib.cpp
@@ -40,7 +40,7 @@ QString formatSize(QWORD qwSize)
 
   if (qwSize < llKiloB) // In Bytes
     {
-      strSize = i18n("1 byte", "%1 bytes").arg(qwSize);
+      strSize = i18n("1 byte", "%1 bytes").arg((unsigned long)qwSize);
     }
   else if (qwSize < llMegaB) // In KiloBytes
     {
